// circabc-frontend/src/app/core/generated/circabc/model/userProfile.ts
/**
 * CIRCABC REST API
 * This is the first version of the CIRCABC REST API used by the new User Interface 
 *
 * The version of the OpenAPI document: 4.2.0
 * Contact: DIGIT-CIRCABC-SUPPORT@ec.europa.eu
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { User } from './user';
import { Profile } from './profile';


/**
 * the representation of the membership of a user inside an Interest group
 */
export interface UserProfile { 
    user?: User;
    profile?: Profile;
    expirationDate?: string;
}