// circabc-frontend/src/app/core/generated/circabc/model/userPostDefinition.ts
/**
 * CIRCABC REST API
 * This is the first version of the CIRCABC REST API used by the new User Interface 
 *
 * The version of the OpenAPI document: 4.2.0
 * Contact: DIGIT-CIRCABC-SUPPORT@ec.europa.eu
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * a new CIRCABC user
 */
export interface UserPostDefinition { 
    userId?: string;
    firstname?: string;
    lastname?: string;
    email?: string;
    phone?: string;
    title?: string;
    companyId?: string;
    fax?: string;
    urlAddress?: string;
    postalAddress?: string;
    description?: string;
    password?: string;
    currentIgId?: string;
}